FPSCounter is an Entity built with FPS and TextSprite components.
TextSprite is a TextureSprite that carries a text attribute and change its
sprite to match its text attribute.
TODO: find folders for other platformes than Windows.
TODO: change colors, size, etc...
TODO: allow the TextSprite to use SoftwareSprite.

FPS must keep a reference to its TextSprite in order to change it when the
callback is triggered after one second. Is there a better way to do this?